Ability to not follow a process
This introduces an ability to not-follow a process in the `proc` widget.
I  having to do `gg` whenever I visit `ytop` to find out the process
which is currently using high mem/CPU. It'd be great to have the ability
to not always track the currently highlighted process.

This patch is a POC for that. What it does is:

 - Introduce a new flag on the widget: `follow_proc`, which is false by
   default. If it is set to `true`, the `render` will always jump to the
   position for the highlighted position. If it is `false`, it'll keep
   the selected position the same, no matter which process lands on it.
 - When ever the user does a scroll action (except `gg`), `ytop` will
   keep track of the process. `gg` or scroll to top on the other hand,
   stops the process tracking facilitating a view such as: the current
   process taking the most resources.
 - Also, bind `Esc` to set `follow_proc` as false and disable following
   of a process.

I understand this cannot be merged as is, but I'd like to start a
discussion. If this makes sense, I'd like suggestions on what can be
changed in the patch to merge it in.